requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
Django==3.1.2
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-js-asset==1.2.2
django-taggit==1.5.1
idna==2.10
Pillow==8.0.1
psycopg2==2.8.5
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.4
